refactor(parser): move files
