+++ src/lib/parser/function-comment/regex.ts
@@ -1,4 +1,4 @@
-export default class FunctionCommentParser {
+export default class FunctionCommentRegexParser {
   public keyword: string;
 
   constructor(keyword: string) {
@@ -14,6 +14,26 @@ export default class FunctionCommentParser {
     }
     return false;
   }
+
+  postProcess(line: string): Map<string, string> | null {
+    // initialize map
+    const match = line.match(/#\s+[\w]+[\s\w\(\)\:\*]+/)
+    if (match) {
+      if (line.startsWith("#")) {
+        const result = new Map<string, string>();
+        result.set('name', 'name');
+        result.set('type', 'type');
+        result.set('description', 'description');
+        return result
+    }
+    }
+    return null;
+
+
+    
+    
+  }
+
 
   isEndScope(line: string): boolean {
     const result = line.match(/#\s?(\w+\s?\w+)/);
